Add unit tests for repository and view model and CI workflow
